requirements.txt
mysql-connector-python
fastapi
uvicorn
argon2-cffi
python-dotenv